requirements.txt
google-genai>=1.0.0
python-dotenv>=1.0.0
pyaudio>=0.2.14
pyobjc-framework-Cocoa>=10.0
pyobjc-framework-ApplicationServices>=10.0
pyobjc-framework-Quartz>=10.0
playwright>=1.40.0
websockets>=12.0
pywebview>=5.0
